+++ app/data.py
@@ -1,12 +1,14 @@
 from types import SimpleNamespace
 from typing import List
 from config_local import Config
+from app.csv_sniff import sniff_delimiter
 import json
 import io
 import gzip
 import re
 import csv
 import os
+import tempfile
 
 if hasattr(Config, 'S3'):
     import boto3
@@ -23,6 +25,23 @@ def to_alnum(string):
     """Get rid of non alpahunmeric characters except underscores
     """
     return ''.join(char for char in string if char.isalnum() or char == '_')
+
+
+def xsv_to_array2d(in_xsv: str) -> List[List[str]]:
+    # read max 10 lines and join them to make a sample for the dialect sniffer
+    xsv_stream = io.StringIO(in_xsv)
+    xsv_stream.seek(0)
+    lines = []
+    for i in range(10):
+        try:
+            lines.append(next(xsv_stream))
+        except StopIteration:
+            pass
+    delimiter = sniff_delimiter(''.join(lines))
+    # read CSV
+    xsv_stream.seek(0)
+    reader = csv.reader(xsv_stream, delimiter=delimiter)
+    return list(reader)
 
 
 def array2d_to_csv(in_list: List[List[str]]) -> io.StringIO:
@@ -42,8 +61,8 @@ def generate_table_stmt(schema, table, columns, text_type_name='VARCHAR'):
     return f'CREATE TABLE {schema}.{table}({cols_type})'
 
 
-def s3_copy(bucket: str, key: str, csv_reader):
-    """copy a pandas csv_reader as a csv.gz to s3
+def s3_copy(bucket: str, key: str, data: List[List[str]]):
+    """copy array data as a csv.gz to s3
     """
     def compress(string, cp='utf-8'):
         out = io.BytesIO()
@@ -51,10 +70,11 @@ def s3_copy(bucket: str, key: str, csv_reader):
             f.write(string.encode(cp))
         return out.getvalue()
 
-    header = next(csv_reader)
+    iterator = iter(data)
+    header = next(iterator)
     body_io = io.StringIO()
     writer = csv.writer(body_io)
-    _ = list(writer.writerow(row) for row in csv_reader)
+    writer.writerows(iterator)
     body_io.seek(0)
     body = compress(body_io.read())
     s3client.put_object(Bucket=bucket, Key=key, Body=body)
@@ -64,20 +84,19 @@ def sqlify(name: str):
     return re.sub(r"[^a-zA-Z0-9]+", "_", name.lower())
 
 
-def destination_redshift(csv_data: io.StringIO, table_name: str, path: str):
+def destination_redshift(data: List[List[str]], table_name: str, path: str):
     import sqlalchemy
     from sqlalchemy.sql import text
 
     dsn = 'redshift://{user}:{password}@{host}:{port}/{dbname}'.format(**Config.DB)
     engine = sqlalchemy.create_engine(dsn)
 
-    reader = csv.reader(csv_data)
     key = f'excel-to-database/{table_name}.csv.gz'
     arn = Config.S3['arn']
     bucket = Config.S3['bucket']
 
     # load to s3 bucket
-    s3_copy(bucket, key, reader)
+    s3_copy(bucket, key, data)
 
     # load to redshift
     schema_name = 'x_excel'
@@ -98,9 +117,8 @@ def destination_redshift(csv_data: io.StringIO, table_name: str, path: str):
     col_names = [col_name.values()[0].lower() for col_name in col_names]
 
     # compare sorted and lower col names as it is tricky to control case and order (not an ideal solution)
-    csv_data.seek(0)
-    header = list(to_alnum(col).lower() for col in next(reader))
-    n_records = len(list(reader))
+    header = list(to_alnum(col).lower() for col in data[0])
+    n_records = len(data) - 1
     if sorted(col_names) == sorted(header):
         # truncate if cols seem to be the same (will fail if only column order is changed)
         action = 'Truncated'
@@ -115,7 +133,7 @@ def destination_redshift(csv_data: io.StringIO, table_name: str, path: str):
     return f'{action} and loaded into {schema_name}.{table_name}.\n{n_records} records loaded successfully.\n'
 
 
-def destination_local(csv_data: io.StringIO, table_name: str, path: str):
+def destination_local(data: List[List[str]], table_name: str, path: str):
     if path:
         if path[0] != '/':
             path = Config.LOCAL_DEST + '/' + path
@@ -126,18 +144,15 @@ def destination_local(csv_data: io.StringIO, table_name: str, path: str):
         os.makedirs(path)
 
     filename = f'{path}/{table_name}.csv'
-    n_records = 0
+    n_records = len(data) - 1
     with open(filename, 'w') as fp:
         writer = csv.writer(fp)
-        reader = csv.reader(csv_data)
-        for row in reader:
-            writer.writerow(row)
-            n_records += 1
+        writer.writerows(data)
 
-    return f'Created {filename}.\n{n_records-1} records loaded successfully.\n'
+    return f'Created {filename}.\n{n_records} records loaded successfully.\n'
 
 
-def destination_azuredw(csv_data: io.StringIO, table_name: str, path: str):
+def destination_azuredw(data: List[List[str]], table_name: str, path: str):
     import pyodbc
 
     conn = pyodbc.connect(
@@ -148,11 +163,9 @@ def destination_azuredw(csv_data: io.StringIO, table_name: str, path: str):
     conn.autocommit = True
     cursor = conn.cursor()
 
-    reader = csv.reader(csv_data)
-
     # compare sorted and upper col names as it is tricky to control case and order (not an ideal solution)
-    csv_data.seek(0)
-    header = next(reader)
+    iterator = iter(data)
+    header = next(data)
 
     schema_name = 'x_excel'
     if path:
@@ -172,9 +185,8 @@ def destination_azuredw(csv_data: io.StringIO, table_name: str, path: str):
     cursor.execute(generate_table_stmt(schema_name, table_name, header, 'NVARCHAR(2000)'))
 
     inserts = []
-    n_records = 0
-    for row in reader:
-        n_records += 1
+    n_records = len(data)
+    for row in iterator:
         inserts.append(f"INSERT INTO {schema_name}.{table_name} VALUES (N'" +
                        "', N'".join(map(lambda col: col.replace("'", "''"), row)) +
                        "');\n")
@@ -187,3 +199,46 @@ def destination_azuredw(csv_data: io.StringIO, table_name: str, path: str):
         cursor.execute(statement)
 
     return f'Loaded into {schema_name}.{table_name}.\n{n_records} records loaded successfully.\n'
+
+
+def destination_snowflake(data: List[List[str]], table_name: str, path: str):
+    import snowflake.connector
+    schema_name = 'x_excel'
+
+    with snowflake.connector.connect(**Config.DB) as connection, tempfile.TemporaryDirectory() as tmpdir:
+        col_names = connection.execute(f'SELECT COLUMN_NAME FROM INFORMATION_SCHEMA.COLUMNS WHERE lower(TABLE_NAME)=\'{table_name}\'')
+        col_names = [col_name.values()[0].lower() for col_name in col_names]
+
+        # compare sorted and lower col names as it is tricky to control case and order (not an ideal solution)
+        iterator = iter(data)
+        header = list(to_alnum(col).lower() for col in next(iterat0r))
+        n_records = len(data)
+        if n_records:
+            gzfilename = f"{tmpdir}/x_excel_data.csv.gz"
+            with gzip.open(gzfilename, 'wt') as fp:
+                writer = csv.writer(fp)
+                writer.writerows(iterator)
+
+            connection.execute(f'CREATE SCHEMA IF NOT EXISTS {schema_name}')
+
+            if sorted(col_names) == sorted(header):
+                # truncate if cols seem to be the same (will fail if only column order is changed)
+                action = 'Truncated'
+                connection.execute(f'TRUNCATE TABLE {schema_name}.{table_name}')
+            else:
+                # drop if col names not the same
+                action = 'Dropped'
+                connection.execute(f'DROP TABLE IF EXISTS {schema_name}.{table_name} CASCADE')
+                connection.execute(generate_table_stmt(schema_name, table_name, header, text_type_name="TEXT"))
+
+            connection.execute(f'USE SCHEMA {schema_name}')
+            connection.cursor().execute(f'CREATE OR REPLACE TEMPORARY STAGE {table_name};')
+            connection.cursor().execute(f'PUT file://{gzfilename} @%{table_name}')
+            connection.cursor().execute(f'''
+                COPY INTO {table_name} file_format = (
+                    TYPE = CSV
+                    FIELD_OPTIONALLY_ENCLOSED_BY = '"'
+                    ERROR_ON_COLUMN_COUNT_MISMATCH = FALSE
+                )'''
+            )
+    return f'{action} and loaded into {schema_name}.{table_name}.\n{n_records} records loaded successfully.\n'